+++ packages/cli/src/commands/epochs/finish.test.ts
@@ -2,7 +2,7 @@ import { newKitFromWeb3 } from '@celo/contractkit'
 import { testWithAnvilL2 } from '@celo/dev-utils/lib/anvil-test'
 import { timeTravel } from '@celo/dev-utils/lib/ganache-test'
 import BigNumber from 'bignumber.js'
-import { testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
+import { stripAnsiCodesFromNestedArray, testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
 import Finish from './finish'
 import Start from './start'
 
@@ -10,18 +10,21 @@ process.env.NO_SYNCCHECK = 'true'
 
 testWithAnvilL2('epochs:finish cmd', (web3) => {
   it('Warns when epoch process is not yet started', async () => {
+    const logMock = jest.spyOn(console, 'log')
     const kit = newKitFromWeb3(web3)
     const accounts = await kit.web3.eth.getAccounts()
     const epochManagerWrapper = await kit.contracts.getEpochManager()
 
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(4)
-    const res = await testLocallyWithWeb3Node(Finish, ['--from', accounts[0]], web3)
-
-    expect(res).toEqual('Epoch process is not started yet')
+    await expect(
+      testLocallyWithWeb3Node(Finish, ['--from', accounts[0]], web3)
+    ).rejects.toMatchInlineSnapshot(`[Error: Epoch process is not started yet]`)
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(4)
+    expect(stripAnsiCodesFromNestedArray(logMock.mock.calls)).toMatchInlineSnapshot(`[]`)
   })
 
   it('finishes epoch process successfully', async () => {
+    const logMock = jest.spyOn(console, 'log')
     const kit = newKitFromWeb3(web3)
     const accounts = await kit.web3.eth.getAccounts()
     const epochManagerWrapper = await kit.contracts.getEpochManager()
@@ -38,5 +41,21 @@ testWithAnvilL2('epochs:finish cmd', (web3) => {
 
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(5)
     expect(await epochManagerWrapper.isTimeForNextEpoch()).toEqual(false)
+    expect(stripAnsiCodesFromNestedArray(logMock.mock.calls)).toMatchInlineSnapshot(`
+      [
+        [
+          "SendTransaction: startNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+        [
+          "SendTransaction: finishNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+      ]
+    `)
   })
 })