+++ packages/cli/src/commands/epochs/switch.test.ts
@@ -2,7 +2,7 @@ import { newKitFromWeb3 } from '@celo/contractkit'
 import { testWithAnvilL2 } from '@celo/dev-utils/lib/anvil-test'
 import { timeTravel } from '@celo/dev-utils/lib/ganache-test'
 import BigNumber from 'bignumber.js'
-import { testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
+import { stripAnsiCodesFromNestedArray, testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
 import Start from './start'
 import Switch from './switch'
 
@@ -10,18 +10,21 @@ process.env.NO_SYNCCHECK = 'true'
 
 testWithAnvilL2('epochs:switch cmd', (web3) => {
   it('Warns only when next epoch is not due when switching', async () => {
+    const logMock = jest.spyOn(console, 'log')
     const kit = newKitFromWeb3(web3)
     const accounts = await kit.web3.eth.getAccounts()
     const epochManagerWrapper = await kit.contracts.getEpochManager()
 
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(4)
-    const res = await testLocallyWithWeb3Node(Switch, ['--from', accounts[0]], web3)
-
-    expect(res).toEqual('It is not time for the next epoch yet')
+    await expect(
+      testLocallyWithWeb3Node(Switch, ['--from', accounts[0]], web3)
+    ).rejects.toMatchInlineSnapshot(`[Error: It is not time for the next epoch yet]`)
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(4)
+    expect(stripAnsiCodesFromNestedArray(logMock.mock.calls)).toMatchInlineSnapshot(`[]`)
   })
 
   it('switches epoch successfully', async () => {
+    const logMock = jest.spyOn(console, 'log')
     const kit = newKitFromWeb3(web3)
     const accounts = await kit.web3.eth.getAccounts()
     const epochManagerWrapper = await kit.contracts.getEpochManager()
@@ -36,9 +39,26 @@ testWithAnvilL2('epochs:switch cmd', (web3) => {
 
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(5)
     expect(await epochManagerWrapper.isTimeForNextEpoch()).toEqual(false)
+    expect(stripAnsiCodesFromNestedArray(logMock.mock.calls)).toMatchInlineSnapshot(`
+      [
+        [
+          "SendTransaction: startNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+        [
+          "SendTransaction: finishNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+      ]
+    `)
   })
 
   it('switches epoch successfully which already has started process', async () => {
+    const logMock = jest.spyOn(console, 'log')
     const kit = newKitFromWeb3(web3)
     const accounts = await kit.web3.eth.getAccounts()
     const epochManagerWrapper = await kit.contracts.getEpochManager()
@@ -54,5 +74,21 @@ testWithAnvilL2('epochs:switch cmd', (web3) => {
 
     expect(await epochManagerWrapper.getCurrentEpochNumber()).toEqual(5)
     expect(await epochManagerWrapper.isTimeForNextEpoch()).toEqual(false)
+    expect(stripAnsiCodesFromNestedArray(logMock.mock.calls)).toMatchInlineSnapshot(`
+      [
+        [
+          "SendTransaction: startNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+        [
+          "SendTransaction: finishNextEpoch",
+        ],
+        [
+          "txHash: 0xtxhash",
+        ],
+      ]
+    `)
   })
 })