+++ packages/cli/src/commands/lockedcelo/show.ts
@@ -6,7 +6,9 @@ import { ViewCommmandFlags } from '../../utils/flags'
 
 export default class Show extends BaseCommand {
   static description =
-    'Show Locked Gold information for a given account. This includes the total amount of locked celo, the amount being used for voting in Validator Elections, the Locked Gold balance this account is required to maintain due to a registered Validator or Validator Group, and any pending withdrawals that have been initiated via "lockedgold:unlock".'
+    'Show Locked Celo information for a given account. This includes the total amount of locked celo, the amount being used for voting in Validator Elections, the Locked Celo balance this account is required to maintain due to a registered Validator or Validator Group, and any pending withdrawals that have been initiated via "lockedcelo:unlock".'
+
+  static aliases = ['lockedgold:show']
 
   static flags = {
     ...ViewCommmandFlags,
@@ -22,10 +24,10 @@ export default class Show extends BaseCommand {
     const kit = await this.getKit()
     const { args } = await this.parse(Show)
     const account = args.arg1 as string
-    const lockedGold = await kit.contracts.getLockedGold()
+    const lockedCELO = await kit.contracts.getLockedGold()
 
     await newCheckBuilder(this).isAccount(account).runChecks()
 
-    printValueMapRecursive(await lockedGold.getAccountSummary(account))
+    printValueMapRecursive(await lockedCELO.getAccountSummary(account))
   }
 }