+++ packages/cli/src/commands/network/parameters.test.ts
@@ -61,7 +61,7 @@ testWithAnvilL1('network:parameters', (web3: Web3) => {
         stageDurations: 
           Execution: 1 week
           Referendum: 1 day
-      LockedGold: 
+      LockedCelo: 
         totalLockedGold: 120000000000000000000000 (~1.200e+23)
         unlockingPeriod: 6 hours
       Reserve: