+++ packages/cli/src/commands/validator/list.test.ts
@@ -4,7 +4,7 @@ import { ux } from '@oclif/core'
 import Web3 from 'web3'
 import { stripAnsiCodesFromNestedArray, testLocallyWithWeb3Node } from '../../test-utils/cliUtils'
 import Register from '../account/register'
-import Lock from '../lockedgold/lock'
+import Lock from '../lockedcelo/lock'
 import ListValidators from './list'
 import ValidatorRegister from './register'