+++ packages/sdk/contractkit/src/contract-cache.test.ts
@@ -10,7 +10,7 @@ const TestedWrappers: ValidWrappers[] = [
   CeloContract.StableToken,
   CeloContract.StableTokenEUR,
   CeloContract.Validators,
-  CeloContract.LockedGold,
+  CeloContract.LockedCelo,
 ]
 
 function newWrapperCache() {
@@ -41,6 +41,33 @@ describe('getContract()', () => {
     const contract1 = await contractCache.getContract(CeloContract.MultiSig, address1)
     const contract2 = await contractCache.getContract(CeloContract.MultiSig, address2)
     expect(contract1?.address).not.toEqual(contract2?.address)
+  })
+  describe('get contract methods', () => {
+    const exclusionList = new Set([
+      'StableTokenEUR',
+      'StableTokenBRL',
+      'Registry',
+      'Random',
+      'UniswapFeeHandlerSeller',
+      'MentoFeeHandlerSeller',
+      'GovernanceSlasher',
+      'FeeHandler',
+      'EpochManagerEnabler',
+      'CeloUnreleasedTreasury',
+      'ERC20',
+    ])
+    const eligibleMethods = Object.values(CeloContract).filter((name) => !exclusionList.has(name))
+    for (const contractName of eligibleMethods) {
+      const method = `get${contractName}`
+      it(`has a method ${method}`, async () => {
+        // @ts-ignore
+        expect(contractCache[method]).toBeDefined()
+        // @ts-ignore
+        expect(typeof contractCache[method]).toBe('function')
+        // @ts-ignore
+        expect(contractCache[method]()).resolves.toBeDefined()
+      })
+    }
   })
 })